requirements.txt
pandas
requests
fuzzywuzzy
airflowctl
streamlit
snowflake-connector-python
numpy
matplotlib